+++ instrumentation/opentelemetry-instrumentation-asgi/tests/test_asgi_custom_headers.py
@@ -123,12 +123,8 @@ class TestCustomHeaders(AsgiTestBase, TestBase):
         self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.request.header.custom_test_header_1": (
-                "test-header-value-1",
-            ),
-            "http.request.header.custom_test_header_2": (
-                "test-header-value-2",
-            ),
+            "http.request.header.custom_test_header_1": ("test-header-value-1",),
+            "http.request.header.custom_test_header_2": ("test-header-value-2",),
             "http.request.header.regex_test_header_1": ("Regex Test Value 1",),
             "http.request.header.regex_test_header_2": (
                 "RegexTestValue2,RegexTestValue3",
@@ -169,14 +165,10 @@ class TestCustomHeaders(AsgiTestBase, TestBase):
         self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.request.header.custom_test_header_1": (
-                "test-header-value-1",
-            ),
+            "http.request.header.custom_test_header_1": ("test-header-value-1",),
         }
         not_expected = {
-            "http.request.header.custom_test_header_2": (
-                "test-header-value-2",
-            ),
+            "http.request.header.custom_test_header_2": ("test-header-value-2",),
         }
         for span in span_list:
             if span.kind == SpanKind.SERVER:
@@ -195,12 +187,8 @@ class TestCustomHeaders(AsgiTestBase, TestBase):
         self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.response.header.custom_test_header_1": (
-                "test-header-value-1",
-            ),
-            "http.response.header.custom_test_header_2": (
-                "test-header-value-2",
-            ),
+            "http.response.header.custom_test_header_1": ("test-header-value-1",),
+            "http.response.header.custom_test_header_2": ("test-header-value-2",),
             "http.response.header.my_custom_regex_header_1": (
                 "my-custom-regex-value-1,my-custom-regex-value-2",
             ),
@@ -242,9 +230,7 @@ class TestCustomHeaders(AsgiTestBase, TestBase):
         self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         not_expected = {
-            "http.response.header.custom_test_header_3": (
-                "test-header-value-3",
-            ),
+            "http.response.header.custom_test_header_3": ("test-header-value-3",),
         }
         for span in span_list:
             if span.kind == SpanKind.SERVER:
@@ -276,12 +262,8 @@ class TestCustomHeaders(AsgiTestBase, TestBase):
         self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.request.header.custom_test_header_1": (
-                "test-header-value-1",
-            ),
-            "http.request.header.custom_test_header_2": (
-                "test-header-value-2",
-            ),
+            "http.request.header.custom_test_header_1": ("test-header-value-1",),
+            "http.request.header.custom_test_header_2": ("test-header-value-2",),
             "http.request.header.regex_test_header_1": ("Regex Test Value 1",),
             "http.request.header.regex_test_header_2": (
                 "RegexTestValue2,RegexTestValue3",
@@ -314,9 +296,7 @@ class TestCustomHeaders(AsgiTestBase, TestBase):
         self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         not_expected = {
-            "http.request.header.custom_test_header_3": (
-                "test-header-value-3",
-            ),
+            "http.request.header.custom_test_header_3": ("test-header-value-3",),
         }
         for span in span_list:
             if span.kind == SpanKind.SERVER:
@@ -346,12 +326,8 @@ class TestCustomHeaders(AsgiTestBase, TestBase):
         self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.response.header.custom_test_header_1": (
-                "test-header-value-1",
-            ),
-            "http.response.header.custom_test_header_2": (
-                "test-header-value-2",
-            ),
+            "http.response.header.custom_test_header_1": ("test-header-value-1",),
+            "http.response.header.custom_test_header_2": ("test-header-value-2",),
             "http.response.header.my_custom_regex_header_1": (
                 "my-custom-regex-value-1,my-custom-regex-value-2",
             ),
@@ -387,9 +363,7 @@ class TestCustomHeaders(AsgiTestBase, TestBase):
         self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         not_expected = {
-            "http.response.header.custom_test_header_3": (
-                "test-header-value-3",
-            ),
+            "http.response.header.custom_test_header_3": ("test-header-value-3",),
         }
         for span in span_list:
             if span.kind == SpanKind.SERVER:
@@ -414,27 +388,15 @@ class TestCustomHeadersEnv(TestCustomHeaders):
         super().setUp()
 
     def tearDown(self):
-        os.environ.pop(
-            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS, None
-        )
-        os.environ.pop(
-            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SERVER_REQUEST, None
-        )
-        os.environ.pop(
-            OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SERVER_RESPONSE, None
-        )
+        os.environ.pop(OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SANITIZE_FIELDS, None)
+        os.environ.pop(OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SERVER_REQUEST, None)
+        os.environ.pop(OTEL_INSTRUMENTATION_HTTP_CAPTURE_HEADERS_SERVER_RESPONSE, None)
         super().tearDown()
 
 
 class TestCustomHeadersConstructor(TestCustomHeaders):
     constructor_params = {
-        "http_capture_headers_sanitize_fields": SANITIZE_FIELDS_TEST_VALUE.split(
-            ","
-        ),
-        "http_capture_headers_server_request": SERVER_REQUEST_TEST_VALUE.split(
-            ","
-        ),
-        "http_capture_headers_server_response": SERVER_RESPONSE_TEST_VALUE.split(
-            ","
-        ),
+        "http_capture_headers_sanitize_fields": SANITIZE_FIELDS_TEST_VALUE.split(","),
+        "http_capture_headers_server_request": SERVER_REQUEST_TEST_VALUE.split(","),
+        "http_capture_headers_server_response": SERVER_RESPONSE_TEST_VALUE.split(","),
     }